+++ Localization/deDE.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-07-24 18-32-54 ]] --
+-- [[ Exported at 2022-07-27 07-11-54 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --